更新
更新

+++ src/main/java/com/gionee/baserom/search/job/GetHaokanData.java
@@ -26,28 +26,33 @@ public class GetHaokanData {
 		JSONObject json = JSONObject.parseObject(newsData);
 		JSONArray datas = json.getJSONArray("news_data");
 		
-		//删除旧数据
-		adNewsService.deleteByType(1);
-		
-		int n = 0;
-		for(int i=0;i<3;i++){
-			JSONObject obj = datas.getJSONObject(i);
-			
-			AdNews ad = new AdNews();
-			ad.setName(obj.getString("name"));
-			ad.setLink(obj.getString("link"));
-			ad.setImg(obj.getString("thumb_img"));
-			AdNews existAd = adNewsService.selectByExample(ad);
-			if(existAd == null){
-				ad.setEnable(1);
-				ad.setCreateTime(new Date());
-				ad.setType(1); //0：手动添加的。1：从第三方接口获取的
-				n += adNewsService.addAdNews(ad);
-			}else{
-				logger.info("已经存在好看数据：" + obj.toJSONString());
+		if(datas!=null && datas.size()!=0){
+			//删除旧数据
+			adNewsService.deleteByType(1);
+			int n = 0;
+			for(int i=0;i<datas.size();i++){
+				
+				if(i==3){
+					break;
+				}
+				
+				JSONObject obj = datas.getJSONObject(i);
+				
+				AdNews ad = new AdNews();
+				ad.setName(obj.getString("name"));
+				ad.setLink(obj.getString("link"));
+				ad.setImg(obj.getString("thumb_img"));
+				AdNews existAd = adNewsService.selectByExample(ad);
+				if(existAd == null){
+					ad.setEnable(1);
+					ad.setCreateTime(new Date());
+					ad.setType(1); //0：手动添加的。1：从第三方接口获取的
+					n += adNewsService.addAdNews(ad);
+				}else{
+					logger.info("已经存在好看数据：" + obj.toJSONString());
+				}
 			}
+			logger.info("更新好看数据：" + n + "条。");
 		}
-		logger.info("更新好看数据：" + n + "条。");
-		
 	}
 }
